MAP-2411 Add endpoint to delete draft locations and refactor related logic
