+++ src/main/kotlin/uk/gov/justice/digital/hmpps/locationsinsideprison/jpa/Location.kt
@@ -173,7 +173,7 @@ abstract class Location(
 
   open fun isActive() = status == LocationStatus.ACTIVE && !isPermanentlyDeactivated()
   open fun isArchived() = status == LocationStatus.ARCHIVED
-  private fun isDraft() = status == LocationStatus.DRAFT
+  fun isDraft() = status == LocationStatus.DRAFT
   open fun isLocked() = false
 
   open fun isResidentialRoomOrConvertedCell() = false
@@ -292,11 +292,11 @@ abstract class Location(
     val subLocations = mutableListOf<Location>()
 
     fun traverse(location: Location) {
-      if (this != location) {
-        subLocations.add(location)
-      }
       for (childLocation in location.childLocations) {
         traverse(childLocation)
+      }
+      if (this != location) {
+        subLocations.add(location)
       }
     }
