+++ src/test/kotlin/uk/gov/justice/digital/hmpps/locationsinsideprison/resource/LocationResidentialResourceTest.kt
@@ -7,17 +7,11 @@ import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.springframework.test.json.JsonCompareMode
 import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.Capacity
-import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.CellInitialisationRequest
-import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.CreateEntireWingRequest
 import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.CreateResidentialLocationRequest
-import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.CreateWingAndStructureRequest
 import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.DerivedLocationStatus
 import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.LegacyLocation
-import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.LevelAboveCells
 import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.LocationTest
-import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.NewCellRequest
 import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.PatchResidentialLocationRequest
-import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.ResidentialStructuralType
 import uk.gov.justice.digital.hmpps.locationsinsideprison.dto.UpdateLocationLocalNameRequest
 import uk.gov.justice.digital.hmpps.locationsinsideprison.integration.CommonDataTestBase
 import uk.gov.justice.digital.hmpps.locationsinsideprison.integration.EXPECTED_USERNAME
@@ -455,578 +449,6 @@ class LocationResidentialResourceTest : CommonDataTestBase() {
 """,
             JsonCompareMode.LENIENT,
           )
-      }
-    }
-  }
-
-  @DisplayName("POST /locations/create-cells")
-  @Nested
-  inner class CreateCellsTest {
-
-    fun createLandingRequest(
-      startingCellNumber: Int = 1,
-      numberOfCells: Int = 1,
-      aboveLevelCode: String = "J",
-      parentLocation: UUID? = null,
-      workingCap: Int = 1,
-      cna: Int = 1,
-      specialistCellTypes: Set<SpecialistCellType> = setOf(SpecialistCellType.ACCESSIBLE_CELL),
-      locationType: ResidentialStructuralType = ResidentialStructuralType.LANDING,
-    ) = CellInitialisationRequest(
-      prisonId = "MDI",
-      parentLocation = parentLocation,
-      newLevelAboveCells = LevelAboveCells(
-        levelCode = aboveLevelCode,
-        locationType = locationType,
-        levelLocalName = "$locationType $aboveLevelCode",
-      ),
-      accommodationType = AccommodationType.NORMAL_ACCOMMODATION,
-      cellsUsedFor = setOf(UsedForType.STANDARD_ACCOMMODATION),
-      cells = (1..numberOfCells).map { index ->
-        NewCellRequest(
-          code = "%03d".format(index - 1 + startingCellNumber),
-          cellMark = "$aboveLevelCode-%03d".format(index - 1 + startingCellNumber),
-          maxCapacity = 1,
-          workingCapacity = workingCap,
-          capacityNormalAccommodation = cna,
-          specialistCellTypes = specialistCellTypes,
-          inCellSanitation = true,
-        )
-      }.toSet(),
-    )
-
-    @Nested
-    inner class Validation {
-      @Test
-      fun `access client error bad data`() {
-        webTestClient.post().uri("/locations/create-cells")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue("""{"prisonId": ""}""")
-          .exchange()
-          .expectStatus().is4xxClientError
-      }
-
-      @Test
-      fun `duplicate location is rejected`() {
-        webTestClient.post().uri("/locations/create-cells")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(jsonString(createLandingRequest(aboveLevelCode = landingZ1.getCode(), parentLocation = wingZ.id)))
-          .exchange()
-          .expectStatus().is4xxClientError
-      }
-
-      @Test
-      fun `request without a parent location or a new location is rejected`() {
-        webTestClient.post().uri("/locations/create-cells")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(jsonString(createLandingRequest().copy(newLevelAboveCells = null)))
-          .exchange()
-          .expectStatus().is4xxClientError
-      }
-
-      @Test
-      fun `request with clashing cell locations is rejected`() {
-        webTestClient.post().uri("/locations/create-cells")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(createLandingRequest(parentLocation = landingZ1.id).copy(newLevelAboveCells = null))
-          .exchange()
-          .expectStatus().is4xxClientError
-      }
-
-      @Test
-      fun `request with specialist cells with non zero CNA or Working Capacity are rejected`() {
-        webTestClient.post().uri("/locations/create-cells")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(createLandingRequest(specialistCellTypes = setOf(SpecialistCellType.BIOHAZARD_DIRTY_PROTEST)))
-          .exchange()
-          .expectStatus().is4xxClientError
-      }
-    }
-
-    @Nested
-    inner class HappyPath {
-      @Test
-      fun `can create a landing and a cell`() {
-        val request = createLandingRequest(parentLocation = wingZ.id)
-        webTestClient.post().uri("/locations/create-cells")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(request)
-          .exchange()
-          .expectStatus().isCreated
-          .expectBody().json(
-            // language=json
-            """ 
-             {
-              "prisonId": "${request.prisonId}",
-              "code": "${request.newLevelAboveCells?.levelCode}",
-              "pathHierarchy": "${wingZ.getCode()}-${request.newLevelAboveCells?.levelCode}",
-              "locationType": "LANDING",
-              "status": "DRAFT",
-              "key": "${request.prisonId}-${wingZ.getCode()}-${request.newLevelAboveCells?.levelCode}",
-              "localName": "LANDING J",
-              "accommodationTypes": [
-                "NORMAL_ACCOMMODATION"
-              ],
-              "specialistCellTypes": [
-                "ACCESSIBLE_CELL"
-              ],
-              "usedFor": [
-                "STANDARD_ACCOMMODATION"
-              ],
-              "capacity": {
-                "maxCapacity": 1,
-                "workingCapacity": 0
-              },
-              "certification": {
-                "certified": false,
-                "capacityOfCertifiedCell": 1
-              },
-              "childLocations": [
-                {
-                  "key": "MDI-Z-J-001",
-                  "cellMark": "J-001",
-                  "locationType": "CELL",
-                  "capacity": {
-                    "maxCapacity": 1,
-                    "workingCapacity": 0
-                  },
-                  "oldWorkingCapacity": 1,
-                  "certification": {
-                    "certified": false,
-                    "capacityOfCertifiedCell": 1
-                  },
-                  "accommodationTypes": [
-                    "NORMAL_ACCOMMODATION"
-                  ],
-                  "specialistCellTypes": [
-                    "ACCESSIBLE_CELL"
-                  ],
-                  "usedFor": [
-                    "STANDARD_ACCOMMODATION"
-                  ],
-                  "status": "DRAFT",
-                  "inCellSanitation": true,
-                  "level": 3,
-                  "leafLevel": true
-                }
-              ]
-            }
-          """,
-            JsonCompareMode.LENIENT,
-          )
-
-        assertThat(getNumberOfMessagesCurrentlyOnQueue()).isZero()
-      }
-
-      @Test
-      fun `can create a cell below a top level location`() {
-        val request = createLandingRequest(locationType = ResidentialStructuralType.WING)
-        webTestClient.post().uri("/locations/create-cells")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(request)
-          .exchange()
-          .expectStatus().isCreated
-          .expectBody().json(
-            // language=json
-            """ 
-             {
-              "prisonId": "${request.prisonId}",
-              "code": "${request.newLevelAboveCells?.levelCode}",
-              "pathHierarchy": "${request.newLevelAboveCells?.levelCode}",
-              "locationType": "WING",
-              "status": "DRAFT",
-              "key": "${request.prisonId}-${request.newLevelAboveCells?.levelCode}",
-              "localName": "WING J",
-              "accommodationTypes": [
-                "NORMAL_ACCOMMODATION"
-              ],
-              "specialistCellTypes": [
-                "ACCESSIBLE_CELL"
-              ],
-              "usedFor": [
-                "STANDARD_ACCOMMODATION"
-              ],
-              "capacity": {
-                "maxCapacity": 1,
-                "workingCapacity": 0
-              },
-              "certification": {
-                "certified": false,
-                "capacityOfCertifiedCell": 1
-              },
-              "childLocations": [
-                {
-                  "key": "MDI-J-001",
-                  "cellMark": "J-001",
-                  "locationType": "CELL",
-                  "capacity": {
-                    "maxCapacity": 1,
-                    "workingCapacity": 0
-                  },
-                  "oldWorkingCapacity": 1,
-                  "certification": {
-                    "certified": false,
-                    "capacityOfCertifiedCell": 1
-                  },
-                  "accommodationTypes": [
-                    "NORMAL_ACCOMMODATION"
-                  ],
-                  "specialistCellTypes": [
-                    "ACCESSIBLE_CELL"
-                  ],
-                  "usedFor": [
-                    "STANDARD_ACCOMMODATION"
-                  ],
-                  "status": "DRAFT",
-                  "inCellSanitation": true,
-                  "level": 2,
-                  "leafLevel": true
-                }
-              ]
-            }
-          """,
-            JsonCompareMode.LENIENT,
-          )
-
-        assertThat(getNumberOfMessagesCurrentlyOnQueue()).isZero()
-      }
-
-      @Test
-      fun `can create a cell below an existing level location`() {
-        val request = createLandingRequest(parentLocation = landingZ1.id, startingCellNumber = 10).copy(newLevelAboveCells = null)
-        webTestClient.post().uri("/locations/create-cells")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(request)
-          .exchange()
-          .expectStatus().isCreated
-          .expectBody().json(
-            // language=json
-            """ 
-             {
-              "prisonId": "${request.prisonId}",
-              "code": "${landingZ1.getCode()}",
-              "pathHierarchy": "${landingZ1.getPathHierarchy()}",
-              "locationType": "${landingZ1.locationType}",
-              "status": "ACTIVE",
-              "key": "${landingZ1.getKey()}",
-              "localName": "Landing 1",
-              "accommodationTypes": [
-                "NORMAL_ACCOMMODATION",
-                "CARE_AND_SEPARATION"
-              ],
-              "specialistCellTypes": [
-                "ACCESSIBLE_CELL"
-              ],
-              "usedFor": [
-                "STANDARD_ACCOMMODATION"
-              ],
-              "capacity": {
-                "maxCapacity": 5,
-                "workingCapacity": 4
-              },
-              "certification": {
-                "certified": true,
-                "capacityOfCertifiedCell": 5
-              },
-              "childLocations": [
-                {
-                  "key": "MDI-Z-1-001",
-                  "status": "ACTIVE"
-                },
-                {
-                  "key": "MDI-Z-1-002",
-                  "status": "ACTIVE"
-                },
-                {
-                  "key": "MDI-Z-1-01S",
-                  "status": "ACTIVE"
-                },  
-                {
-                  "key": "MDI-Z-1-010",
-                  "status": "DRAFT",
-                  "inCellSanitation": true
-                }  
-              ]
-            }
-          """,
-            JsonCompareMode.LENIENT,
-          )
-
-        assertThat(getNumberOfMessagesCurrentlyOnQueue()).isZero()
-      }
-
-      @Test
-      fun `can create just a landing with no cells`() {
-        val request = createLandingRequest(numberOfCells = 0)
-        webTestClient.post().uri("/locations/create-cells")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(request)
-          .exchange()
-          .expectStatus().isCreated
-          .expectBody().json(
-            // language=json
-            """ 
-             {
-              "prisonId": "${request.prisonId}",
-              "code": "${request.newLevelAboveCells?.levelCode}",
-              "pathHierarchy": "${request.newLevelAboveCells?.levelCode}",
-              "locationType": "LANDING",
-              "status": "DRAFT",
-              "key": "${request.prisonId}-${request.newLevelAboveCells?.levelCode}",
-              "localName": "LANDING J",
-              "accommodationTypes": [],
-              "specialistCellTypes": [],
-              "usedFor": [ ],
-              "capacity": {
-                "maxCapacity": 0,
-                "workingCapacity": 0
-              },
-              "certification": {
-                "certified": false,
-                "capacityOfCertifiedCell": 0
-              },
-              "childLocations": []
-            }
-          """,
-            JsonCompareMode.LENIENT,
-          )
-
-        assertThat(getNumberOfMessagesCurrentlyOnQueue()).isZero()
-      }
-    }
-  }
-
-  @DisplayName("POST /locations/create-wing")
-  @Nested
-  inner class CreateWingTest {
-    var createWingAndStructure = CreateWingAndStructureRequest(
-      prisonId = "MDI",
-      wingCode = "Y",
-      wingDescription = "Y Wing",
-      wingStructure = listOf(ResidentialStructuralType.WING, ResidentialStructuralType.LANDING, ResidentialStructuralType.CELL),
-    )
-
-    @Nested
-    inner class Validation {
-      @Test
-      fun `access client error bad data`() {
-        webTestClient.post().uri("/locations/create-wing")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue("""{"prisonId": ""}""")
-          .exchange()
-          .expectStatus().is4xxClientError
-      }
-
-      @Test
-      fun `duplicate location is rejected`() {
-        webTestClient.post().uri("/locations/create-wing")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(jsonString(createWingAndStructure.copy(wingCode = "Z")))
-          .exchange()
-          .expectStatus().is4xxClientError
-      }
-    }
-
-    @Nested
-    inner class HappyPath {
-      @Test
-      fun `can create a wing with a 3 tier structure`() {
-        webTestClient.post().uri("/locations/create-wing")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(createWingAndStructure)
-          .exchange()
-          .expectStatus().isCreated
-          .expectBody().json(
-            // language=json
-            """ 
-             {
-              "prisonId": "MDI",
-              "code": "Y",
-              "pathHierarchy": "Y",
-              "locationType": "WING",
-              "status":"DRAFT",
-              "key": "MDI-Y",
-              "localName": "Y Wing",
-              "wingStructure": [
-                "WING",
-                "LANDING",
-                "CELL"
-              ],
-              "capacity": {
-                "maxCapacity": 0,
-                "workingCapacity": 0
-              }
-            }
-          """,
-            JsonCompareMode.LENIENT,
-          )
-
-        assertThat(getNumberOfMessagesCurrentlyOnQueue()).isZero()
-      }
-
-      @Test
-      fun `can create a wing with a 4 tier structure`() {
-        webTestClient.post().uri("/locations/create-wing")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(createWingAndStructure.copy(wingStructure = listOf(ResidentialStructuralType.WING, ResidentialStructuralType.LANDING, ResidentialStructuralType.SPUR, ResidentialStructuralType.CELL)))
-          .exchange()
-          .expectStatus().isCreated
-          .expectBody().json(
-            // language=json
-            """ 
-             {
-              "prisonId": "MDI",
-              "code": "Y",
-              "pathHierarchy": "Y",
-              "locationType": "WING",
-              "status":"DRAFT",
-              "key": "MDI-Y",
-              "localName": "Y Wing",
-              "wingStructure": [
-                "WING",
-                "LANDING",
-                "SPUR",
-                "CELL"
-              ],
-              "capacity": {
-                "maxCapacity": 0,
-                "workingCapacity": 0
-              }
-            }
-          """,
-            JsonCompareMode.LENIENT,
-          )
-
-        assertThat(getNumberOfMessagesCurrentlyOnQueue()).isZero()
-      }
-    }
-  }
-
-  @DisplayName("POST /locations/create-entire-wing")
-  @Nested
-  inner class CreateEntireWingTest {
-    var createEntireWingRequest = CreateEntireWingRequest(
-      prisonId = "MDI",
-      wingCode = "Y",
-      wingDescription = "Y Wing",
-      numberOfSpurs = 4,
-      numberOfLandings = 2,
-      numberOfCellsPerSection = 2,
-      defaultCellCapacity = 1,
-    )
-
-    @Nested
-    inner class Validation {
-      @Test
-      fun `access client error bad data`() {
-        webTestClient.post().uri("/locations/create-entire-wing")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue("""{"prisonId": ""}""")
-          .exchange()
-          .expectStatus().is4xxClientError
-      }
-
-      @Test
-      fun `duplicate location is rejected`() {
-        webTestClient.post().uri("/locations/create-entire-wing")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(jsonString(createEntireWingRequest.copy(wingCode = "Z")))
-          .exchange()
-          .expectStatus().is4xxClientError
-      }
-    }
-
-    @Nested
-    inner class HappyPath {
-      @Test
-      fun `can create an entire wing with 4 spurs, 2 landings, and 2 cells per landing`() {
-        webTestClient.post().uri("/locations/create-entire-wing")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(createEntireWingRequest)
-          .exchange()
-          .expectStatus().isCreated
-          .expectBody().json(
-            // language=json
-            """ 
-             {
-              "prisonId": "MDI",
-              "code": "Y",
-              "pathHierarchy": "Y",
-              "locationType": "WING",
-              "active": true,
-              "key": "MDI-Y",
-              "localName": "Y Wing",
-              
-              "capacity": {
-                "maxCapacity": 16,
-                "workingCapacity": 16
-              },
-              "certification": {
-                "certified": true,
-                "capacityOfCertifiedCell": 16
-              }
-            }
-          """,
-            JsonCompareMode.LENIENT,
-          )
-
-        getDomainEvents(16).let {
-          assertThat(it).hasSize(16)
-        }
-      }
-
-      @Test
-      fun `can create an entire wing with 2 landings, 2 cells per landing`() {
-        webTestClient.post().uri("/locations/create-entire-wing")
-          .headers(setAuthorisation(roles = listOf("ROLE_MAINTAIN_LOCATIONS"), scopes = listOf("write")))
-          .header("Content-Type", "application/json")
-          .bodyValue(createEntireWingRequest.copy(wingCode = "X", wingDescription = "X Wing", numberOfLandings = 2, numberOfSpurs = 0, numberOfCellsPerSection = 2, defaultCellCapacity = 2))
-          .exchange()
-          .expectStatus().isCreated
-          .expectBody().json(
-            // language=json
-            """ 
-             {
-              "prisonId": "MDI",
-              "code": "X",
-              "pathHierarchy": "X",
-              "locationType": "WING",
-              "active": true,
-              "key": "MDI-X",
-              "localName": "X Wing",
-              
-              "capacity": {
-                "maxCapacity": 8,
-                "workingCapacity": 8
-              },
-              "certification": {
-                "certified": true,
-                "capacityOfCertifiedCell": 8
-              }
-            }
-          """,
-            JsonCompareMode.LENIENT,
-          )
-
-        getDomainEvents(4).let {
-          assertThat(it).hasSize(4)
-        }
       }
     }
   }
